Relocate route data tests
